some cleanup of the new tests, fixing comments
